requirements.txt
faker
confluent-kafka
neo4j
chromadb
sentence-transformers
streamlit
langchain
langchain-google-genai
langchain-community
python-dotenv